api/requirements.txt
azure-functions
azure-cosmos
azure-identity
azure-iot-hub
